Migrate E2EE room action
